* Revert "POSIX: Get render_process_host to build." * Revert "Build fix: release builds seemed to break"
Review URL: http://codereview.chromium.org/18460


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@8386 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/renderer_host/render_process_host.cc
@@ -13,11 +13,10 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace {
 
 unsigned int GetMaxRendererProcessCount() {
-  // Defines the maximum number of renderer processes according to the
-  // amount of installed memory as reported by the OS. The table
-  // values are calculated by assuming that you want the renderers to
-  // use half of the installed ram and assuming that each tab uses
-  // ~25MB.
+  // Defines the maximum number of renderer processes according to the amount
+  // of installed memory as reported by the OS. The table values are calculated
+  // by assuming that you want the renderers to use half of the installed ram
+  // and assuming that each tab uses ~25MB.
   static const int kMaxRenderersByRamTier[] = {
     4,                                  // less than 256MB
     8,                                  // 256MB
@@ -27,7 +26,7 @@ unsigned int GetMaxRendererProcessCount() {
 
   static unsigned int max_count = 0;
   if (!max_count) {
-    size_t memory_tier = base::SysInfo::AmountOfPhysicalMemoryMB() / 256;
+    int memory_tier = base::SysInfo::AmountOfPhysicalMemoryMB() / 256;
     if (memory_tier >= arraysize(kMaxRenderersByRamTier))
       max_count = chrome::kMaxRendererProcessCount;
     else
